+++ src/WEDMap/WED_MapBkgnd.h
@@ -32,10 +32,10 @@ public:
 						 WED_MapBkgnd(GUI_Pane * host, WED_MapZoomerNew * zoomer, IResolver * resolver);
 	virtual				~WED_MapBkgnd();
 
-	virtual	void		DrawVisualization(int inCurrent, GUI_GraphState * g);
-	virtual	void		DrawStructure	 (int inCurrent, GUI_GraphState * g);
+	virtual	void		DrawVisualization(intptr_t inCurrent, GUI_GraphState * g);
+	virtual	void		DrawStructure	 (intptr_t inCurrent, GUI_GraphState * g);
 
-	virtual	void		GetCaps(int& draw_ent_v, int& draw_ent_s, int& cares_about_sel);
+	virtual	void		GetCaps(intptr_t& draw_ent_v, intptr_t& draw_ent_s, intptr_t& cares_about_sel);
 		
 };
