+++ src/WEDMap/WED_MapLayer.h
@@ -37,18 +37,18 @@ public:
 	virtual				~WED_MapLayer();
 
 	// These provide generalized drawing routines.  Use this to draw background images and other such stuff.
-	virtual	void		DrawVisualization		(int inCurrent, GUI_GraphState * g)	{ }
-	virtual	void		DrawStructure			(int inCurrent, GUI_GraphState * g) { }
-	virtual	void		DrawSelected			(int inCurrent, GUI_GraphState * g) { }
+	virtual	void		DrawVisualization		(intptr_t inCurrent, GUI_GraphState * g)	{ }
+	virtual	void		DrawStructure			(intptr_t inCurrent, GUI_GraphState * g) { }
+	virtual	void		DrawSelected			(intptr_t inCurrent, GUI_GraphState * g) { }
 	
 	// These draw specific entities.  Use these to draw pieces of the data model.  Only visible entities will be passed in!	
-	virtual	bool		DrawEntityVisualization	(int inCurrent, IGISEntity * entity, GUI_GraphState * g, int selected) { return false; }
-	virtual	bool		DrawEntityStructure		(int inCurrent, IGISEntity * entity, GUI_GraphState * g, int selected) { return false; }
+	virtual	bool		DrawEntityVisualization	(intptr_t inCurrent, IGISEntity * entity, GUI_GraphState * g, int selected) { return false; }
+	virtual	bool		DrawEntityStructure		(intptr_t inCurrent, IGISEntity * entity, GUI_GraphState * g, int selected) { return false; }
 
 	// Extra iterations over the entity hiearchy get very expensive.  This routine returns whether a layer wants 
 	// per-entity drawing passes for either structure or visualization.  We can also say whether we need "seleted" to be
 	// calculated accurately - checking selection slows down the sped of drawing passes.
-	virtual	void		GetCaps(int& draw_ent_v, int& draw_ent_s, int& cares_about_sel)=0;
+	virtual	void		GetCaps(intptr_t& draw_ent_v, intptr_t& draw_ent_s, intptr_t& cares_about_sel)=0;
 
 protected:
