+++ src/WEDMap/WED_StructureLayer.h
@@ -17,9 +17,9 @@ public:
 			bool		GetVerticesShowing(void) const;
 			void		SetVerticesShowing(bool show);
 
-	virtual	bool		DrawEntityStructure		(int inCurrent, IGISEntity * entity, GUI_GraphState * g, int selected);
-	virtual	void		DrawStructure			(int inCurrent, GUI_GraphState * g);
-	virtual	void		GetCaps(int& draw_ent_v, int& draw_ent_s, int& cares_about_sel);
+	virtual	bool		DrawEntityStructure		(intptr_t inCurrent, IGISEntity * entity, GUI_GraphState * g, int selected);
+	virtual	void		DrawStructure			(intptr_t inCurrent, GUI_GraphState * g);
+	virtual	void		GetCaps(intptr_t& draw_ent_v, intptr_t& draw_ent_s, intptr_t& cares_about_sel);
 
 private:
